+++ HPXMLtoOpenStudio/measure.rb
@@ -1474,11 +1474,11 @@ class OSModel
 
         airloop_map[sys_id] = HVAC.apply_ground_to_air_heat_pump(model, runner, weather, heat_pump,
                                                                  sequential_heat_load_fracs, sequential_cool_load_fracs,
-                                                                 conditioned_zone, @hpxml.site.ground_conductivity, @hvac_unavailable_periods)
-
+                                                                 conditioned_zone, @hpxml.site.ground_conductivity, @hpxml.site.ground_diffusivity,
+                                                                 @hvac_unavailable_periods)
       end
 
-      next unless not heat_pump.backup_system.nil?
+      next if heat_pump.backup_system.nil?
 
       equipment_list = model.getZoneHVACEquipmentLists.find { |el| el.thermalZone == conditioned_zone }